Add files via upload
The video tutorials page would be a new page to add to the Dear PyGui Github. It contains all videos by the core developers and a video by the community. For some users, videos are a good way to get an idea about what a package does before installing and a different way to learn about using the package. 

Initially, I though about adding a short description to each video, but ended up summarising and annotation the videos. Like always, feel free to let me know what you think or change it however you like.

-- discord_testpilot