Move python build to dedicated directory
This makes development in git-enabled IDEs easier, as the build
artifacts will be ignored by cpython's .gitignore and thus won't show up
in the IDE's unstaged changes. It also resolves this for the git CLI.

This is only relevant on Mac and Linux. The build artifacts on Windows
were already ignored in the same way.