Get linux build fully working (#883)
* Fix wheel build to work with non-debug python

* Move python build to dedicated directory

This makes development in git-enabled IDEs easier, as the build
artifacts will be ignored by cpython's .gitignore and thus won't show up
in the IDE's unstaged changes. It also resolves this for the git CLI.

This is only relevant on Mac and Linux. The build artifacts on Windows
were already ignored in the same way.

* Make all scripts executable

git didn't know the execute bit was set on these files, so the user had
to set them manually before building.

* Stop forcing rebuild of core when building wheel

* Use compiled version for building wheel

This is especially important on Debian/Ubuntu, because on those and
other distros with apt, ensurepip is disabled and there's no portable
way to install pip that doesn't have assumptions about package names.

* Parse and pass jobs argument correctly

Before, it was never reached by getopts and thus the python compilation
would only ever run on a single core.

* Testing CI fix

* Update dpg_sources.cmake

* Update dpg_sources.cmake

Co-authored-by: Jonathan Hoffstadt <jonathanhoffstadt@yahoo.com>