Update ImGuiFileDialog.cpp
The bug manifests as a freeze or crash of DPG when attempting to access a directory that the current user does not have permission to access on windows using file_dialog. This commit fixes the issue by adding a code snippet that checks the permission to access the target directory before attempting to do so. If the user does not have permission, the program will not access the target directory and stay in the current path.